+++ api/docs/CHANGELOG.md
@@ -17,6 +17,7 @@ keywords: "API, Docker, rcli, REST, documentation"
 
 [Docker Engine API v1.52](https://docs.docker.com/reference/api/engine/version/v1.52/) documentation
 
+* `POST /exec/{id}/signal` is a new endpoint that can be used for sending a signal to an exec instance.
 * `GET /images/{name}/get` now accepts multiple `platform` query-arguments
   to allow selecting which platform(s) of a multi-platform image must be
   saved.